[frontend] fix(security-coverage): fix xolor on mitre widget
